requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.3.2               # via kombu
appnope==0.1.0            # via ipython
backcall==0.1.0           # via ipython
billiard==3.5.0.5         # via celery
boto3==1.9.69             # via ocw-data-parser
botocore==1.12.69         # via boto3, s3transfer
celery==4.2.0
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
decorator==4.3.0          # via ipython, traitlets
dj-database-url==0.4.2
django-redis==4.7.0
django-server-status==0.4.0
django-webpack-loader==0.5.0
django==2.0.8
djangorestframework==3.9.0
docutils==0.14            # via botocore
elasticsearch==6.3.1      # via django-server-status
factory-boy==2.11.1
faker==1.0.1
idna==2.8                 # via requests
ipython-genutils==0.2.0   # via traitlets
ipython==7.2.0
jedi==0.13.1              # via ipython
jmespath==0.9.3           # via boto3, botocore
kombu==4.2.2              # via celery, django-server-status
newrelic==4.8.0.110
ocw-data-parser==0.0.3
parso==0.3.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
prompt-toolkit==2.0.7     # via ipython
psycopg2==2.7.3.2
ptyprocess==0.6.0         # via pexpect
pygments==2.3.0           # via ipython
python-dateutil==2.7.5    # via botocore, faker
pytz==2018.7              # via celery, django
raven==6.9.0
redis==2.10.6
requests==2.21.0
s3transfer==0.1.13        # via boto3
six==1.12.0               # via django-server-status, faker, prompt-toolkit, python-dateutil, traitlets
text-unidecode==1.2       # via faker
traitlets==4.3.2          # via ipython
urllib3==1.24.1           # via botocore, elasticsearch, requests
uwsgi==2.0.17.1
vine==1.1.4               # via amqp
wcwidth==0.1.7            # via prompt-toolkit
